fa-info: do a mechanical refactor of ent_info to prepare it to be split up
We introduce a function append, which replaces the concatenations.  We drop the extra
spaces.  Then we pass it through localise_cat_table.

This increases the size of the code and is annoying to read, but it lets child functions
return localised strings and also avoids them having to worry about where spaces should go.  Soon, we will divide
the mega-function into pieces, and then those pieces can directly return localised strings.

Since we have to debug this anyway, it seems apropos to make the big change now: let's not do
two passes of testing, when we can just do one as we fix 2.0.