Introduce the console command /fac, the same as /c but accessible and in the Factorio Access environment
If you are sighted /c is fine. If you are blind /c is wrap things in print and then
go check the launcher output, and probably also throw serpent around that.  /fac instead
handles figuring out the common cases of when to use serpent, overrides lua print to work with the launcher, and lets
the specified code run in the FA environment (e.g. it can see storage).

Crashes are guarded with pcall on a best effort basis, but there is probably still a way to crash this.  Still, it's for devs not players.